Get PNG product name from header and merge vectors
It is no longer the case that each group's main product will
be called "reduced", so it is necessary to determine the
product name for the preview images from their headers rather
than assuming that it will always be "reduced".

Also updated the PNG image merging subroutine to deal
with vector plots, which should be composited with
the representative images.