Merge filter-specific PNGs to make color-composite
The merge_ukirt_pngs subroutine has had two compositing subroutines
extracted from it.  The first, merge_ukirt_png_vec_bg, is the
existing code to overlay a picture with its vector plot.  The
second, merge_ukirt_png_multicolor, is a new routine to
sort the images into wavelength order and merge them into a
color-composite.  It can handle 2 images (making a blue and red
composite) or 3 images (blue/green/red composite).  If given
more than 3 it takes only the first 3 (by wavelength).

Also the merge_ukirt_pngs subroutine no longer dies if it
finds a combination of thumbnails which it cannot merge.
A warning is printed instead.  This is to allow processing
to proceed as the preview isn't absolutely essential
to the process.