Update provenance prior to conversion to FITS
The convert_ukirt_sdfs subroutine now calls
JSA::Starlink::prov_update_parent_path on each file before converting
it to FITS.  To make this straightforward, the loop which finalizes
the filenames and converts the files has been separated into two parts.
Therefore the complete set of filename mappings is available when
updating the provenance, and can be used to determine which files
will be in the archive and with what new file names.